# OpenAPI/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/activities:
    get:
      tags:
      - Activities
      summary: Get activities with pagination
      description: Retrieve activities with optional filtering and pagination
      operationId: getActivitiesPaged
      parameters:
      - name: fetchAll
        in: query
        description: Fetch all activities without pagination
        required: true
        schema:
          type: boolean
      - name: from
        in: query
        description: "Filter by source (user, mural, etc)"
        required: false
        schema:
          type: string
      - name: id
        in: query
        description: Filter by source ID
        required: false
        schema:
          type: integer
          format: int64
      - name: startPage
        in: query
        description: Starting page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: pagesToFetch
        in: query
        description: Number of pages to fetch
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        description: Number of entries per page
        required: false
        schema:
          type: integer
          format: int32
      - name: cond
        in: query
        description: Filter condition
        required: false
        schema:
          type: string
      - name: visibility
        in: query
        description: Filter by visibility
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Activities retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDTO"
    post:
      tags:
      - Activities
      summary: Create new activity
      description: Create a new activity from an uploaded GPX file
      operationId: createActivity
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  description: GPX file to upload
                  format: binary
      responses:
        "200":
          description: Activity created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "400":
          description: Invalid file format
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
  /api/activities/{activityId}/route:
    post:
      tags:
      - Activities
      summary: Add route to activity
      description: Associate a route with an activity
      operationId: addRoute
      parameters:
      - name: activityId
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int64
        required: true
      responses:
        "200":
          description: Route added successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "404":
          description: Activity or route not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
...
  /api/routes/{id}/visibility:
    patch:
      tags:
      - Routes
      summary: Change route visibility
      description: "Update the visibility settings of a route. Request body should\
        \ contain 'visibility' (PUBLIC, PRIVATE, MURAL_PUBLIC, MURAL_SPECIFIC) and\
        \ optionally 'allowedMuralsList' as a comma-separated string in brackets for\
        \ MURAL_SPECIFIC visibility"
      operationId: changeVisibility
      parameters:
      - name: id
        in: path
        description: Route ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "404":
          description: Route not found
          content:
            '*/*':
              schema:
                type: boolean
        "204":
          description: Visibility changed successfully
          content:
            '*/*':
              schema:
                type: boolean
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: boolean
...
  /api/activities/{id}:
    get:
      tags:
      - Activities
      summary: Get activity by ID
      description: Retrieve an activity's information by its ID
      operationId: getActivity
      parameters:
      - name: id
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      - name: mural
        in: query
        description: Optional mural ID
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Activity found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "404":
          description: Activity not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
    delete:
      tags:
      - Activities
      summary: Delete activity
      description: Delete an activity by its ID
      operationId: deleteActivity
      parameters:
      - name: id
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Activity deleted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "404":
          description: Activity not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
    patch:
      tags:
      - Activities
      summary: Edit activity
      description: Update activity details
      operationId: editActivity
      parameters:
      - name: id
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityEditDTO"
        required: true
      responses:
        "404":
          description: Activity not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "200":
          description: Activity updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
  /api/activities/{id}/visibility:
    patch:
      tags:
      - Activities
      summary: Change activity visibility
      description: "Update the visibility settings of an activity. Request body should\
        \ contain 'visibility' (PUBLIC, PRIVATE, MURAL_PUBLIC, MURAL_SPECIFIC) and\
        \ optionally 'allowedMuralsList' as a comma-separated string in brackets for\
        \ MURAL_SPECIFIC visibility"
      operationId: changeVisibility_1
      parameters:
      - name: id
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "404":
          description: Activity not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "200":
          description: Visibility changed successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
  /api/activities/visibility/mural:
    patch:
      tags:
      - Activities
      summary: Hide activities from mural
      description: Make selected activities not visible to a specific mural
      operationId: makeActivitiesNotVisibleToMural
      parameters:
      - name: id
        in: query
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
        required: true
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: Activities visibility updated
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
...
  /api/activities/byIds:
    get:
      tags:
      - Activities
      summary: Get multiple activities by IDs
      description: Retrieve multiple activities by their IDs. Returns only activities
        the user has access to; check ATRA-requested-forbidden header to see if any
        requested activities were filtered
      operationId: getActivitiesByIds
      parameters:
      - name: ids
        in: query
        description: List of activity IDs
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int64
      - name: muralId
        in: query
        description: Optional mural ID
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Activities retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDTO"
  /api/activities/OwnedInMural:
    get:
      tags:
      - Activities
      summary: Get owned activities in mural
      description: Retrieve activities owned by the current user in a specific mural
      operationId: getOwnedActivitiesInMural
      parameters:
      - name: muralId
        in: query
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDTO"
        "200":
          description: Activities retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityDTO"
...
  /api/activities/{id}/route:
    delete:
      tags:
      - Activities
      summary: Remove route from activity
      description: Remove an associated route from an activity
      operationId: removeRoute
      parameters:
      - name: id
        in: path
        description: Activity ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Activity not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "200":
          description: Route removed successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ActivityDTO"
components:
  schemas:
...
    ActivitySummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        startTime:
          type: string
          format: date-time
        totalDistance:
          type: number
          format: double
        totalTime:
          type: integer
          format: int64
        elevationGain:
          type: number
          format: double
        averages:
          type: object
          additionalProperties:
            type: number
            format: double
        records:
          type: object
          additionalProperties:
            type: string
        upRecords:
          type: object
          additionalProperties:
            type: string
          writeOnly: true
        upAverages:
          type: object
          additionalProperties:
            type: number
            format: double
          writeOnly: true
...
    NamedId:
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          format: int64
...
    ActivityDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        startTime:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/NamedId"
        route:
          $ref: "#/components/schemas/NamedId"
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/DataPoint"
        streams:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        summary:
          $ref: "#/components/schemas/ActivitySummary"
        visibility:
          type: string
          enum:
          - PRIVATE
          - MURAL_SPECIFIC
          - MURAL_PUBLIC
          - PUBLIC
        allowedMurals:
          type: array
          items:
            type: integer
            format: int64
    DataPoint:
      type: object
      properties:
        time:
          type: string
          format: date-time
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        ele:
          type: number
          format: double
        hr:
          type: integer
          format: int32
        cad:
          type: integer
          format: int32
        other:
          type: object
          additionalProperties:
            type: string
...
    ActivityEditDTO:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
...